[Fix #744] Prerelease improvements (#745)
* [Fix #744] Suspend/resume status and associated event

Workflow suspend and resume event should be sent regardless a task was
in fact suspended or not. Same applies to workflow status.

Signed-off-by: Francisco Javier Tirado Sarti <ftirados@redhat.com>

* [Fix #744] Add api dependency to test and example

Although already included because jackson-impl depends on it, it is a
good practise to explicitly add API as dependency because this
dependency is an implementation detail and the usage of the reader API
by the exampple is not.

Signed-off-by: Francisco Javier Tirado Sarti <ftirados@redhat.com>

* [Fix #744] Disabling event life cycle publishing

In some scenarios, users wont need the cloud events to be published.
This allows user to disable publishing at application level.

Signed-off-by: Francisco Javier Tirado Sarti <ftirados@redhat.com>

* [Fix #744] Improving EventPublisher/EventConsumer

There might be more that one EventPublisher

EventPublishers are either added programatically or through service
loader.

There will be only one consumer. If user want to handler more than one
consuming event broker, he should provide an specific implementation
that deals with them.

The one consumer is added programatically (priortiy) or default to
service loader.

If not eventConsumer was provides, InMemoryEvents is used as event
consumer and publisher.

ChatBotIT and LifeCycleEventTest has been changed accordingly.

Signed-off-by: Francisco Javier Tirado Sarti <ftirados@redhat.com>

* [Fix #744] EmitExecutor shorter version

Signed-off-by: Francisco Javier Tirado Sarti <ftirados@redhat.com>

---------

Signed-off-by: Francisco Javier Tirado Sarti <ftirados@redhat.com>